ai-part/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
scikit-learn
python-dotenv